==> processing_pipelines/perturb_spatial/06_PCA.ipynb <==
import scanpy as sc
import os
import numpy as np
import matplotlib.pyplot as plt

Adding paths to the perturb experiment

data_dir = "/mnt/sata2/Analysis_Alex_2/perturb1"
exper = "SI_2"

adata = sc.read(os.path.join(data_dir, exper, "adatas", "04_tissue_cleared.h5ad"))

Names of the guide RNA genes and the guide RNAs that they correspond to

guide_rna_genes = ["Muc5ac", "Neurog3", "Fer1l6"]
corresponding_guide_rnas = ["sgCd19", "sgThy1", "sgCxcr3"]

Calculate PCA without the guide RNA expression

guide_rna_indices = np.where(~adata.var.index.isin(guide_rna_genes))[0]
pca = sc.tl.pca(adata.X[:, guide_rna_indices], random_state=0)

adata.obsm["X_pca"] = pca

adata.write(os.path.join(data_dir, exper, "adatas", "05_reference_mapped.h5ad"))
