==> processing_pipelines/perturb_spatial/09_tokenizing_datasets_for_celltyping.ipynb <==
Tokenizing cells for Geneformer, which can be used to predict cell types in the perturb data from the time course data

import sys

sys.path.append("/home/amonell/Geneformer")
from geneformer import TranscriptomeTokenizer
import numpy as np
import os
import pandas as pd

Put the path to the reference data

import scanpy as sc

ad = sc.read("timecourse.h5ad")

data_dir = "/mnt/sata2/Analysis_Alex_2/perturb1/SI*"

##### Converting gene names to human homolog ensembl ids
We provide the following file in the current directory

ensembl = pd.read_csv("gProfiler_hsapiens_9-18-2023_2-31-49 PM.csv").drop_duplicates(
    "initial_alias"
)

ad.var["ensembl_id"] = [i for i in ensembl["converted_alias"] if not pd.isna(i)][
    : len(ad.var.index)
]
ad.var.index = ad.var.ensembl_id.tolist()
ad.obs["n_counts"] = np.array(np.sum(ad.X, axis=1)).flatten()
ad.obs["organ_major"] = "SI"

ad.obs["input_ids"] = [i for i in range(len(ad.obs.index.tolist()))]
d = {}
for i in ad.var.index.tolist():
    d[i] = True

tk = TranscriptomeTokenizer(
    {"Subtype": "cell_type", "organ_major": "organ_major"}, nproc=4
)

Tokenize datasets to annotate

geneformer_path = "/mnt/sata1/Analysis_Alex/Geneformer"

import glob

ensembl = pd.read_csv("gProfiler_hsapiens_9-18-2023_2-31-49 PM.csv").drop_duplicates(
    "initial_alias"
)

for filename in glob.glob(data_dir):
    outname = os.path.basename(filename)
    fname = os.path.dirname(filename)
    try:
        os.mkdir(os.path.join(geneformer_path, "loom_" + outname))
        ad = sc.read(
            os.path.join(
                fname,
                outname,
                "adatas/05_reference_mapped.h5ad",
            )
        )
        ad.var["ensembl_id"] = [
            i for i in ensembl["converted_alias"] if not pd.isna(i)
        ][: len(ad.var.index)]
        ad.var.index = ad.var.ensembl_id.tolist()
        ad.obs["n_counts"] = np.array(np.sum(ad.X, axis=1)).flatten()
        ad.obs["organ_major"] = "SI"
        ad.obs["Subtype"] = 0
        ad.write_loom(
            os.path.join(geneformer_path, "loom_" + outname + "/" + outname + ".loom")
        )

        tk = TranscriptomeTokenizer(
            {"Subtype": "cell_type", "organ_major": "organ_major"}, nproc=4
        )

        tk.tokenize_data(
            os.path.join(geneformer_path, "loom_" + outname),
            os.path.join(geneformer_path, "loom_" + outname, "tokenized"),
            "train_" + outname,
        )
    except:
        print(outname)
