==> processing_pipelines/perturb_spatial/12_Creating_Final_Object.ipynb <==
import scanpy as sc
import os
import numpy as np
from tqdm.notebook import tqdm
import glob
from scipy.spatial import KDTree
import matplotlib.pyplot as plt
import scvi

input_folders = glob.glob("/mnt/sata2/Analysis_Alex_2/perturb1/SI*")

adatas = []
for input_file in input_folders:
    adatas_ = sc.read(os.path.join(input_file, "adatas", "10_axes_defined.h5ad"))
    adatas_.obs["batch"] = input_file.split("/")[-1]
    adatas.append(adatas_)
adatas = sc.concat(adatas)

adatas.obs = adatas.obs.drop(
    adatas.obs.columns[adatas.obs.columns.str.contains("Topic ")], axis=1
)

Write out combined object to an output folder

output_folder = r"/mnt/sata2/Analysis_Alex_2/perturb1"

try:
    os.makedirs(output_folder)
except:
    pass

adatas.write(os.path.join(output_folder, "integrated_celltyped.h5ad"))

concatenated_all = sc.read(os.path.join(output_folder, "integrated_celltyped.h5ad"))

We also want to clip our epithelial values at 0.6 for visualization

concatenated_all.obs["epithelial_distance_clipped"] = np.clip(
    concatenated_all.obs["epithelial_distance"].values, 0, 0.6
)

Write out the final object

concatenated_all.write(os.path.join(output_folder, "final_celltyped_and_axes.h5ad"))
